rewrote controller constraints javadoc and fixed a few bugs
